Move killLeft into commands.ts
